requirements.txt
Flask==3.1.1
gunicorn==23.0.0
pdfplumber==0.11.6
python-docx==1.1.2
